Moved config to ./config
