+++ config/client.js
@@ -1,0 +1,19 @@
+export default {
+  apiEndpoint: 'http://localhost:3000/api/v1',
+  staticEndpoint: 'http://localhost:3000/static',
+  ctfTitle: ' | yourCTF',
+  ctfName: 'yourCTF',
+  logoUrl: 'https://webite.com/logo.png',
+  sponsors: [
+    {
+      name: 'Big Sponsor',
+      icon: 'https://website.com/logo.png',
+      description: 'Brief description'
+    },
+    {
+      name: 'Smaller Sponsor',
+      description: 'Brief description',
+      small: true
+    }
+  ]
+}
