+++ server/api/auth/register.js
@@ -3,7 +3,7 @@ const emailValidator = require('email-validator')
 const cache = require('../../cache')
 const util = require('../../util')
 const auth = require('../../auth')
-const config = require('../../../config')
+const config = require('../../../config/server')
 const { responses } = require('../../responses')
 
 module.exports = {
