+++ server/api/submitflag.js
@@ -1,7 +1,7 @@
 const db = require('../database')
 const challenges = require('../challenges')
 const { responses } = require('../responses')
-const config = require('../../config')
+const config = require('../../config/server')
 const util = require('../util')
 
 const uuidv4 = require('uuid/v4')
