+++ server/cache/client.js
@@ -1,5 +1,5 @@
 const redis = require('redis')
-const config = require('../../config')
+const config = require('../../config/server')
 
 const client = redis.createClient({
   url: config.redisUrl
