+++ server/database/db.js
@@ -1,4 +1,4 @@
-const config = require('../../config')
+const config = require('../../config/server')
 
 const { Pool } = require('pg')
 const pool = new Pool({
