+++ test/util.js
@@ -1,7 +1,7 @@
 require('ava')
 
 const uuidv4 = require('uuid/v4')
-const config = require('../config')
+const config = require('../config/server')
 
 module.exports = {
   generateTestUser: () => {
